Revised file structure and broke down App.js into multiple components.
